% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_custom_covs}
\alias{add_custom_covs}
\title{Append lavaan syntax with group-specific covariances}
\usage{
add_custom_covs(model, group, data, cov, focal.groups)
}
\arguments{
\item{model}{character, lavaan syntax model}

\item{group}{character, grouping variable}

\item{data}{data frame}

\item{cov}{character, covariance to add, e.g. "variable1 ~~ variable2"}

\item{focal.groups}{Character vector for the groups to add the cov.}
}
\description{
Append lavaan syntax with group-specific covariances
}
\examples{
cov.model <-  "F =~ v1 + v2 v3 + v4 + v5"
cov.model.custom.covs <-
   lav.mod \%>\%
     add_custom_covs("country", dat1,
                     "v1 ~~ v3", c("China", "Indonesia")) \%>\%

     add_custom_covs("country.f", dat1,
                     "v2 ~~ v3", c("Israel"))

}
