% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{clusteredMI}
\alias{clusteredMI}
\title{Run clustered measurement invariance tests}
\usage{
clusteredMI(
  model,
  group,
  data,
  clusters,
  parameters = c("loadings", "intercepts"),
  ref = "configural",
  additional = c("scalar"),
  ...
)
}
\arguments{
\item{model}{character, lavaan syntax model}

\item{group}{character, grouping variable}

\item{data}{data frame}

\item{clusters}{A list of character vectors of the group names to create clusters.}

\item{parameters}{character vector, "all", "loadings", "thresholds", or "intercepts". Clusters are applied to this subset of parameters.}
}
\description{
Run clustered measurement invariance tests
}
\examples{
clusteredMI("F =~ v1 + v2 + v3 + v4", 
         group = "country", 
         data = Dat1, 
         clusters = list(North = c("Norway", "Denmark", "Finland"), 
                         South = c("Spain", "Portugal", "Italy")
                         )
          )

}
