% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.functions.R
\name{compute_correlation}
\alias{compute_correlation}
\title{Computes and formats Fisher's transformed correlation matrices}
\usage{
compute_correlation(data, group)
}
\arguments{
\item{data}{Individual data}

\item{group}{Character. Grouping variable}
}
\description{
Computes and formats Fisher's transformed correlation matrices
}
