% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.functions.R
\name{compute_covariance}
\alias{compute_covariance}
\title{Computes and formats covariance matrices}
\usage{
compute_covariance(data, group)
}
\arguments{
\item{data}{Data containing group variable}

\item{group}{Character, group variable}
}
\description{
Computes and formats covariance matrices
}
