% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.functions.R
\name{dmacs.unsigned}
\alias{dmacs.unsigned}
\title{Dmacs}
\usage{
dmacs.unsigned(LambdaR, ThreshR, LambdaF, ThreshF, MeanF, VarF, SD)
}
\arguments{
\item{lav.model}{fitted lavaan model, configural}
}
\description{
Simply a wrapper arounnd dmacs::lavaan_dmacs() function
}
