% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractAlignment.R
\name{extractAlfDiff}
\alias{extractAlfDiff}
\title{Extracts DIFF alignment results from Mplus output string}
\usage{
extractAlfDiff(input, file = T)
}
\arguments{
\item{input}{Mplus output either as a single string or filename}

\item{file}{Logical. If TRUE, `input` is treated as a filename.}
}
\description{
Extracts DIFF alignment results from Mplus output string
}
\details{
Mplus input file has to include the 'align;' in the OUTPUT section and    'DIFF()' statements in the 'Model prior' section.
}
\examples{
# b.string = readLines("fixed.out")
# alf.out = getAlfDiff(b.string)
}
