% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractAlignment.R
\name{extractAlignment}
\alias{extractAlignment}
\title{Extracts summaries from Mplus alignment output file}
\usage{
extractAlignment(file = "fixed.out", nice.tables = FALSE, silent = FALSE)
}
\arguments{
\item{file}{filename of the Mplus alignment output file}

\item{nice.tables}{Logical. If tables should be send to RStudio Viewer pane using `kable` and `kableExtra` packages.}

\item{silent}{Logical. Used for debugging.}
}
\value{
A list of summary tables.
}
\description{
Extracts summaries from Mplus alignment output file
}
\examples{
\dontrun{  
   align.summ <- extractAlignment("fixed.out") 
   }
}
\seealso{
\code{\link[MIE]{runAlignment}}  and \code{\link[MIE]{extractAlignmentSim}}
}
