% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractDiffOutput.R
\name{extractBayesDiff}
\alias{extractBayesDiff}
\title{Extracts summaries from Mplus Bayesian approximate invariance output file}
\usage{
extractBayesDiff(file)
}
\arguments{
\item{file}{filename of the Mplus alignment output file}
}
\value{
A summary table and a parsed difference outpout to data frame.
}
\description{
Extracts summaries from Mplus Bayesian approximate invariance output file
}
\examples{
\dontrun{  
   align.summ <- extractBayesDiff("BSEM_prior01.out") 
   }
}
\seealso{
\code{\link[MIE]{extractAlignment}}  and \code{\link[MIE]{extractAlignmentSim}}
}
