% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurementInvarianceMplus.R
\name{extractMImplus}
\alias{extractMImplus}
\title{Extract Mplus measurement invariance output}
\usage{
extractMImplus(outfile)
}
\arguments{
\item{outfile}{Character. Path to Mplus output file.}
}
\value{
A summary table with fit indices and their differences between models.
}
\description{
Extract Mplus measurement invariance output
}
\details{
Mplus output file has to contain results of measurement invariance testing, that is, block `ANALYSIS: model= configural metric scalar;`.
}
