% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.functions.R
\name{get_pairs}
\alias{get_pairs}
\title{Extract attribute 'pairs.of.groups'}
\usage{
get_pairs(df)
}
\arguments{
\item{df}{Any object with an attribute "pairs.of.groups"}
}
\description{
Extract attribute 'pairs.of.groups'
}
