% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.functions.R
\name{getMetrics}
\alias{getMetrics}
\title{Computes various metrics of measurement invariance distances}
\usage{
getMetrics(
  data,
  group,
  model,
  ...,
  metric = c("covariance", "correlation", "fit", "parameters")
)
}
\arguments{
\item{data}{Individual data}

\item{group}{Character. Grouping variable}

\item{model}{lavaan model syntax}

\item{...}{Additional arguments passed to \code{\link[MIE]{incrementalFit}} or \code{\link[MIE]{mgcfaParameters}}}

\item{metric}{Character. Metric to be computed. One of "covariance", "correlation", "fit", or "parameters".}
}
\value{
Depending on the value of `metric` argument, returns:
\itemize{
\item "covariance": covariance matrices across groups (output of \code{\link[MIE]{computeCovariance}})
\item "correlation": correlation matrices across groups (output of \code{\link[MIE]{computeCorrelation}})
\item "fit": pairwise decreases in fit indices (output of \code{\link[MIE]{incrementalFit}})
\item "parameters": pairwise differences in parameters (output of \code{\link[MIE]{mgcfaParameters}})
}
}
\description{
Computes various metrics of measurement invariance distances
}
