% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.functions.R
\name{globalMIH}
\alias{globalMIH}
\title{Test for measurement invariance for Higher-Order Factor models}
\usage{
globalMIH(
  model = model,
  ...,
  chi.sq = FALSE,
  omit = "",
  what = c("cfi", "tli", "rmsea", "srmr", "chisq"),
  partial = NULL,
  more = NULL
)
}
\arguments{
\item{model}{lavaan syntax model definition. Keep it simple as it will be appended by the function. Please use one of the loadings to identify the model at each factor order.}

\item{...}{Formula, group, and all the other eligible arguments of \code{\link[lavaan]{cfa}} function.}

\item{chi.sq}{Logical, if chi-square should be reported.}

\item{omit}{Character (vector). Which model should be omitted. Possible values "configural", "metric", "scalar".}

\item{what}{Character (vector) of fit measures to report (see \code{\link[lavaan]{fitMeasures}}}

\item{partial}{List of parameters to release constraints, see `group.partial` of \code{\link[lavOptions]{cfa}}}
}
\description{
Computes configural, multiple metric, and multiple scalar invariance models and compares them.
}
\details{
For convenience, does not follow Rudnev et al. (2018) but instead the following series of constraints: 
\enumerate{
  \item Configural All loadings are free
  \item Metric-1 Loadings for the first-order factors are constrained to equality, first-order intercepts are free, the other means/intercepts are fixed to 0.
  \item Metric-2, 3 etc. Applies the factor loading equality constraints to second-, third-, etc order factors.
  \item Scalar-1 Constrains item intercepts to equality, while freely estimating the first-order factor means/intercepts.
  \item Scalar-2, 3, etc. Sets the means/intercepts of first-order, second-order, etc.  factors free while fixing to 0 all the other means/intercepts (except for the lowest level item intercepts)
}
}
