% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{groupwiseMI}
\alias{groupwiseMI}
\title{Run the same cfa within every group and extract fit indices}
\usage{
groupwiseMI(model, data, group, ..., out = c("fit", "models"))
}
\arguments{
\item{model}{Model in lavaan syntax}

\item{data}{The data}

\item{group}{Character. Grouping variable.}

\item{...}{Other arguments passed to lavaan::cfa}

\item{out}{Character. "fit" returns some fit indices for every group; "models" list of entire fittted models.}
}
\description{
Run the same cfa within every group and extract fit indices
}
