% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.functions.R
\name{incrementalFit}
\alias{incrementalFit}
\title{Run pairwise models and compute decrease in fit}
\usage{
incrementalFit(..., level = "metric")
}
\arguments{
\item{...}{The arguments passed to \code{\link[MIE]{pairwiseFit}}. Required arguments are \code{'model'}, \code{'data'}, and \code{'group'}.}

\item{level}{Character. A model set to be computed. The function will compute a decrease in fit between two models:
\describe{
\item{metric}{(default) between configural and metric models}
\item{scalar}{between metric and scalar models}
\item{at.once}{between configural and scalar models, }
\item{intercepts.first}{between configural and model with equal intercepts/thresholds, but free loadings}
}}
}
\value{
Returns a list with detailed output on every available fit index, and a large matrix used for plotting with \code{\link[MIE]{plotDistances}}
}
\description{
Runs two MGCFA models for each possible pair of groups and computes change in fit.
}
