% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.functions.R
\name{MI_global}
\alias{MI_global}
\title{Test for measureent invariance in three steps}
\usage{
MI_global(...)
}
\arguments{
\item{...}{Formula, group, and all the other eligible arguments of `lavaan::cfa` function.}
}
\description{
Computes configural, metric, and scala invariance models and compares them.
}
