% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{n_invariant}
\alias{n_invariant}
\title{Get the number of groups which show invariance with each of the groups}
\usage{
n_invariant(fit, sort = F, fit.index = "cfi", cutoff = 0.01, drop = NA)
}
\arguments{
\item{fit}{Output of \code{\link{incrementalFit}}.}

\item{sort}{Logical. Should the results be sorted?}

\item{fit.index}{Character. What measure should be used?}

\item{cutoff}{The cutoff of the difference in `fit.index` to decide whether invariance holds.}

\item{drop}{Character list of the groups that should be excluded from calculations.}
}
\value{
Returns a data frame with two columns: group name and the number of groups with invariant parameters.
}
\description{
Get the number of groups which show invariance with each of the groups
}
