% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{n_invariant_matrix}
\alias{n_invariant_matrix}
\title{Get the tree of}
\usage{
n_invariant_matrix(...)
}
\arguments{
\item{...}{See \code{\link{n_invariant}} except `sort` and `drop` arguments.}
}
\value{
Returns a data frame with N-1 columns, where N is number of groups.
}
\description{
Get the tree of
}
\details{
Runs \code{\link{n_invariant}} iteratively, by dropping the least invariant group at each iteration until there are 2 groups left.
}
