% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{nInvariant}
\alias{nInvariant}
\title{Get the number of groups which show invariance with each of the groups}
\usage{
nInvariant(
  fit,
  sort = F,
  fit.index = "cfi",
  cutoff = 0.01,
  drop = NA,
  df = FALSE
)
}
\arguments{
\item{fit}{Output of \code{\link{incrementalFit}}.}

\item{sort}{Logical. Should the results be sorted?}

\item{fit.index}{Character. What measure should be used?}

\item{cutoff}{The cutoff of the difference in `fit.index` to decide whether invariance holds.}

\item{drop}{Character list of the groups that should be excluded from calculations.}

\item{df}{Logical, if the data.frame of the comparison tree should be returned. If TRUE, `sort` and `drop` are ignored.}
}
\value{
If df = FALSE, returns a data frame with two columns: group name and the number of groups with invariant parameters. IF df = TRUE, calculates the number of gorups iteratively, by dropping the least invariant group at each iteration until there are 2 groups left. Returns a data frame with N-1 columns, where N is number of groups.
}
\description{
Get the number of groups which show invariance with each of the groups
}
