% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.functions.R
\name{pairs_of_groups}
\alias{pairs_of_groups}
\title{Produce all possible pairs of groups}
\usage{
pairs_of_groups(variable)
}
\arguments{
\item{variable}{Grouping variable to find the set of all possible unique pairs of values.}
}
\description{
Produce all possible pairs of groups
}
