% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.functions.R
\name{plotDistances}
\alias{plotDistances}
\title{Plot distances using computed measures of distance}
\usage{
plotDistances(measures, n.clusters = "auto", fit.index = "cfi",
  drop = NULL, dist.method = NULL)
}
\arguments{
\item{measures}{Can be result of `MGCFAparameters`, `computeCovariance`, `computeCorrelation`, `incrementalFit`.}

\item{n.clusters}{Number of clusters.}

\item{fit.index}{Index to be used to in represneting measurement invariance distances. Only if the `measures` argument is an output of `incrementalFit`. Can be anything returned by `lavaan::fitMeasures`}

\item{drop}{Vector of group names to be dropped from the plot.}
}
\value{
Computes distances and performs multidimensional scaling (two-dimensional projection). Returns ggplot-based plot.
}
\description{
Plot distances using computed measures of distance
}
