% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAlignment.R
\name{runAlignment}
\alias{runAlignment}
\title{Run alignment procedure using Mplus}
\usage{
runAlignment(
  model,
  group,
  dat,
  categorical = NULL,
  sim.samples = c(100, 500, 1000),
  sim.reps = 500,
  Mplus_com = "mplus",
  path = getwd(),
  summaries = FALSE,
  estimator = "MLR",
  listwise = "OFF",
  processors = 2
)
}
\arguments{
\item{model}{Character. Formula in Mplus format, e.g. "Factor1 BY item1 item2 item3 item4; item4 WITH item3;", see example.}

\item{group}{Character, name of the grouping variable.}

\item{dat}{Data frame containing data.}

\item{categorical}{Character vector of variable names. Indicators that are binary or ordinal. Default is NULL.}

\item{sim.samples}{Vector of integers. Group sample sizes for simulation,  the length of this vector also determines  a number of simulation studies. Default is `c(100, 500, 1000)`. May take a substantial amount of time. Use NULL to avoid running simulations.}

\item{sim.reps}{A number of simulated datasets in each simulation. Default is 500. Use 0 to avoid running simulations.}

\item{Mplus_com}{Sometimes you don't have a direct access to Mplus, so this argument specifies what to send to a system command line. Default value is "mplus".}

\item{path}{Where all the .inp, .out, and .dat files should be stored?}

\item{summaries}{If the \code{\link[MIE]{extractAlignment}}     and \code{\link[MIE]{extractAlignmentSim}}    should be run after all the Mplus work is done. Default is FALSE.}

\item{estimator}{Could be any estimator supported by Mplus. Default is "MLR".}
}
\description{
Facilitates running frequentist alignmnet procedure in Mplus. It creates Mplus input and data from R data.frame, runs the input using Mplus, and provides its summaries back in R.
}
\details{
The function runs in four steps to facilitate setting up and running alignment:
\enumerate{
    \item Converts data.drame into a dat file compatible with Mplus. Saves it on the disk.
    \item Creates Mplus input file for free alignemnent. Saves it on the disk.
    \item Runs this input using the saved data, which produces an .out file.
    \item Reads in the output, selects the appropriate reference group, creates a new input file for fixed alignment, and runs it.
    \item (Optional) Runs simulations to check the reliability of the fixed alignment results.
    \item (Optional) Using \code{\link[MIE]{extractAlignment}} and/or \code{\link[MIE]{extractAlignmentSim}} reads in the output file of the fixed alignment and summarizies them in a convenient, publishable tables.
}
The sequence of free, fixed alignment and simulations follows recommendations of Muthen & Asparouhov (2014).

All the files created during these steps stay on the disk and can be used independently for reporting. For a detailed tutorial on alignment in general see \url{https://maksimrudnev.com/2019/05/01/alignment-tutorial/}
}
\examples{
\dontrun{ aling.res = runAlignment(model = "Moral1 BY prostit homosex abortion divorce;
Moral2 BY benefits taxes bribes;", 
group = "country",
 dat = wvs.s1,
 Mplus_com = "mplus",
 summaries = T
 )
 }
 @seealso \code{\link[MIE]{extractAlignment}}, \code{\link[MIE]{extractAlignmentSim}}, \code{\link[MIE]{extractAlignmentSim}}, \code{\link[MIE]{measurementInvarianceMplus}}
 
}
