% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAlignment.R
\name{runAlignmentSim}
\alias{runAlignmentSim}
\title{Creates Mplus code for alignment simulations and optionally runs it and returns its summaries back in R.}
\usage{
runAlignmentSim(
  model,
  sim.samples = c(100, 500, 1000),
  sim.reps = 500,
  Mplus_com = "mplus",
  path = getwd(),
  summaries = TRUE
)
}
\arguments{
\item{model}{Path to the fixed alignment Mplus output file.}

\item{sim.samples}{Vector of integers. Group sample sizes for simulation,  the length of this vector also determines  a number of simulation studies. Default is `c(100, 500, 1000)`. May take a substantial amount of time. Use NULL to avoid running simulations.}

\item{sim.reps}{A number of simulated datasets in each simulation. Default is 500. Use 0 to avoid running simulations.}

\item{Mplus_com}{Sometimes you don't have a direct access to Mplus, so this argument specifies what to send to a system command line. Default value is "mplus".}

\item{path}{Where all the .inp, .out, and .dat files should be stored?}

\item{summaries}{If the \code{\link[MIE]{extractAlignmentSim}} should be run after all the Mplus work is done. Default is FALSE.}
}
\description{
Creates Mplus code for alignment simulations and optionally runs it and returns its summaries back in R.
}
