% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.functions.R
\name{runMIE}
\alias{runMIE}
\title{Run Shiny MIE application}
\usage{
runMIE(model = NULL, data = NULL, group = NULL, verbose = FALSE)
}
\arguments{
\item{model}{Character, lavaan formula. Special value "demo" to use simulated dataset and model.}

\item{data}{Data.frame containing only the variables to be used in calculations. May contain only numeric variables. If `group` argument is not specified, the group variable should be the first in the data.}

\item{group}{Character, name of the variable in data.}
}
\description{
Run Shiny MIE application
}
\examples{
\dontrun{ 
 if (interactive()) { 
   runMIE("f =~ a1 + a2 + a3", mydr, "municipality") 
 }
}

}
