% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurementInvarianceMplus.R
\name{runMImplus}
\alias{runMImplus}
\title{Runs exact invariance tests in Mplus}
\usage{
runMImplus(
  model,
  group,
  data,
  categorical = NULL,
  filename = NULL,
  Mplus_com = "mplus"
)
}
\arguments{
\item{model}{Character. Formula in Mplus format, e.g. "Factor1 BY item1 item2 item3 item4; item4 WITH item3;", see example.}

\item{group}{Character, name of the grouping variable.}

\item{data}{Data frame containing data.}

\item{categorical}{Character vector of variable names. Indicators that are binary or ordinal. Default is NULL.}

\item{filename}{Optional custom filename}

\item{Mplus_com}{Sometimes you don't have a direct access to Mplus, so this argument specifies what to send to a system command line. Default value is "mplus".}
}
\description{
Runs exact invariance tests in Mplus
}
\examples{
\dontrun{
   runMImplus(model = 
     "F1 BY F114 F115 F116 F117;
      F2 BY F118 F119  F120  F121 F122  F123;",
      group = "S003",
      }

}
