% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{summary.groupwiseCFA}
\alias{summary.groupwiseCFA}
\title{Summarize a list of CFAs produces by lavaan}
\usage{
\method{summary}{groupwiseCFA}(model.list)
}
\arguments{
\item{model.list}{A list of models in lavaan syntax. Can be an object of class groupwise.CFA}
}
\description{
Summarize a list of CFAs produces by lavaan
}
