% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractAlignmentSim.R
\name{extractAlignmentSim}
\alias{extractAlignmentSim}
\title{Extracts summaries of alignment simulations from Mplus output file}
\usage{
extractAlignmentSim(
  sim.outputs = c("sim500.out", "sim100.out", "sim1000.out"),
  silent = FALSE,
  manual = F
)
}
\arguments{
\item{sim.outputs}{a character vector of file names containing simulation results of alignment ran in Mplus.}

\item{silent}{Logical. Used for debugging.}
}
\value{
Invisibly returns a summary table.
}
\description{
Extracts summaries of alignment simulations from Mplus output file
}
\details{
Best used as part of \code{\link[MIE]{runAlignment}}
}
\examples{
\dontrun{  
     align.sim.summ <- extractAlignmentSim (c("sim500.out", "sim100.out", "sim1000.out")) 
     }

}
\seealso{
\code{\link[MIE]{runAlignment}}  and \code{\link[MIE]{extractAlignment}}
}
