% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.functions.R
\name{globalMI}
\alias{globalMI}
\title{Conventional test of measurement invariance in three steps}
\usage{
globalMI(
  ...,
  chi.sq = FALSE,
  omit = "",
  what = c("cfi", "tli", "rmsea", "srmr", "chisq"),
  partial = NULL,
  more = NULL,
  ho = F
)
}
\arguments{
\item{...}{Formula, group, and all the other eligible arguments of \code{\link[lavaan]{cfa}} function.}

\item{chi.sq}{Logical, if chi-square should be reported.}

\item{omit}{Character (vector). Which model should be omitted. Possible values "configural", "metric", "scalar".}

\item{what}{Character (vector) of fit measures to report (see \code{\link[lavaan]{fitMeasures}}}

\item{partial}{List of parameters to release constraints, see `group.partial` of \code{\link[lavOptions]{cfa}}}

\item{more}{More constraints, currently limited to `residuals` and `means`.}

\item{ho}{Currently defunct, for the higher-order factor models invariance see \code{\link[MIE]{globalMIH}}}
}
\description{
Computes configural, metric, and scala invariance models and compares them.
}
