% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.functions.R
\name{globalMI}
\alias{globalMI}
\title{Test for measureent invariance in three steps}
\usage{
globalMI(
  ...,
  chi.sq = FALSE,
  omit = "",
  what = c("cfi", "tli", "rmsea", "srmr", "chisq"),
  partial = NULL
)
}
\arguments{
\item{...}{Formula, group, and all the other eligible arguments of \code{\link[lavaan]{cfa}} function.}

\item{chi.sq}{Logical, if chi-square should be reported.}

\item{omit}{Character (vector). Which model should be omitted. Possible values "configural", "metric", "scalar".}
}
\description{
Computes configural, metric, and scala invariance models and compares them.
}
