% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{groupwiseCFA}
\alias{groupwiseCFA}
\title{Run the same cfa within every group and extract fit indices}
\usage{
groupwiseCFA(
  model,
  data,
  group,
  ...,
  out = NULL,
  what = c("cfi", "rmsea", "chisq", "mod")
)
}
\arguments{
\item{model}{Model in lavaan syntax}

\item{data}{The data}

\item{group}{Character. Grouping variable.}

\item{...}{Other arguments passed to lavaan::cfa}

\item{out}{Legacy argument, keeps only `fit` or `models` parts of the model.}

\item{what}{Fit indices to print and return. Possible values are any fit measures returned by `fitmeasures()`, a special value `mod` adds a parameter with the largest modification index and its MI value.}
}
\description{
Run the same cfa within every group and extract fit indices
}
