% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.functions.R
\name{incrementalFit}
\alias{incrementalFit}
\title{Run pairwise models and compute decrease in fit}
\usage{
incrementalFit(..., level = "metric")
}
\arguments{
\item{...}{The arguments passed to `pairwiseFit``}

\item{level}{Character. If "metric" (default) the decrease of fit between configural and metric model is computed, if "scalar", metric is compared to scalar model fit.}
}
\value{
Returns a list with detailed output on every available fit index, and a large matrix used for plotting with `plotDistances`
}
\description{
Run pairwise models and compute decrease in fit
}
