% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.functions.R
\name{pairwiseFit}
\alias{pairwiseFit}
\title{Computes MGCFA model for a given pair of groups}
\usage{
pairwiseFit(
  model,
  data,
  group = "cntry",
  constraints = c(""),
  pairs.of.groups = NULL,
  message = "Fitting pairwise lavaan models",
  shiny = FALSE,
  ...
)
}
\arguments{
\item{model}{Model in lavaan syntax}

\item{data}{Data object}

\item{group}{Character. Group variable name.}

\item{constraints}{Can be "", "loadings", or c("loadings", "intercepts").}

\item{pairs.of.groups}{Full list of pairs of groups, used in shiny only.}

\item{message}{Notification in shiny.}

\item{shiny}{If it is executed in shiny environment.}

\item{...}{Arguments passed to lavaan 'cfa' function.}
}
\value{
The function returns matrix of fit indices for multiple group CFA models fitted to each possible pair of groups.
}
\description{
Computes MGCFA model for a given pair of groups
}
\details{
Mostly for internal use  within \code{\link[MIE]{incrementalFit}}
}
