% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.functions.R
\name{plotCutoff}
\alias{plotCutoff}
\title{Plots network based on pairwise fit indices reduced to Chen's cutoffs}
\usage{
plotCutoff(
  measures,
  fit.index = "cfi",
  cutoff = NULL,
  weighted = TRUE,
  drop = NULL,
  shiny = F
)
}
\arguments{
\item{measures}{The result of \code{\link[MIE]{incrementalFit}}}

\item{fit.index}{Index to be used to in representing measurement invariance distances. Only if the `measures` argument is an output of \code{\link[MIE]{incrementalFit}}. Should be "cfi", "rmsea", or "srmr", because only for these indices the cutoffs were suggested by Chen (2007).}

\item{cutoff}{Custom cutoff value used to plot the endges between groups.}

\item{weighted}{Logical. If weighted graph should be created accounting for the size of fit intrement. See \code{\link[igraph]{graph_from_adjacency_matrix}} for details.}

\item{drop}{Character vector of group names to be dropped from the plot.}

\item{shiny}{Logical. If the function is executed in a shiny context.}
}
\description{
Plots network based on pairwise fit indices reduced to Chen's cutoffs
}
\details{
The function extracts a given fit indeces from pairwise fitted MGCFAs, and uses cutoff of .01 to identify edges between groups (nodes), so that the groups for whom  invariance is supported, are connected on the plot. The results are plotted using \code{\link[igraph]{cluster_label_prop}}.
}
\seealso{
\code{\link[MIE]{plotDistances}}
}
