% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{summary.groupwiseCFA}
\alias{summary.groupwiseCFA}
\alias{gsumCFA}
\title{Summarize a list of CFAs produced by lavaan}
\usage{
\method{summary}{groupwiseCFA}(
  model.list,
  what = c("cfi", "rmsea", "srmr"),
  diagnose = T,
  corrs = F,
  mod = F,
  n = F
)

gsumCFA(
  model.list,
  what = c("cfi", "rmsea", "srmr"),
  diagnose = T,
  corrs = F,
  mod = F,
  n = F
)
}
\arguments{
\item{model.list}{A list of models in lavaan syntax. Can be an object of class groupwise.CFA}
}
\description{
Summarize a list of CFAs produced by lavaan
}
